+++ HISTORY.md
@@ -36,6 +36,7 @@
 ### 2021-09-08
 * objects: changed category for "switch to scene" to control
 * blocks, objects, threads: new hat block for "when this scene starts"
+* blocks, threads: changed "switch to scene" to be a stop block
 
 ### 2021-09-07
 * blocks, objects, threads: new change-of-scene event